Register render hooks and livewire in plugin
